+++ src/main/java/io/github/zeroone3010/yahueapi/LightType.java
@@ -4,8 +4,7 @@ package io.github.zeroone3010.yahueapi;
  * See https://developers.meethue.com/documentation/supported-lights for further specifications.
  */
 public enum LightType {
-  /** "ON_OFF" enumeration value will be renamed to "ON_OFF_LIGHT" in the future. */
-  @Deprecated ON_OFF,
+  ON_OFF_LIGHT,
   DIMMABLE,
   COLOR_TEMPERATURE,
   COLOR,
@@ -19,7 +18,7 @@ public enum LightType {
     }
     switch (type.toLowerCase()) {
       case "on/off light":
-        return ON_OFF;
+        return ON_OFF_LIGHT;
       case "dimmable light":
         return DIMMABLE;
       case "color temperature light":